Class Diagram
The GameSystem class is the class that is essentially the central control system that governs the pacing and the function of the game. It will hold an ENUM of gameState which will be able to detect and set the phases for gameplay be it menu screens, pausing, exploration, or combat. We will be able to track the states in which the game is in as to control the pacing as well as to facilitate gameplay.

The Menu class is a subset of the GameSystem class as it directly influences the ENUM gameState which has been detailed and set from the GameSystem class. It has multiple functions, the displayMenu, which as its title implies, displays the main menu in the terminal to which the function getUserChoice is then able to parse the user input and subsequently alter the ENUM to the appropriate gameState and progress the game as such. It holds 3 current options: play, pause, and quit which will all be managed through the appropriate menus that it prints.

The Entity class is the base class that controls all the moving parts and objects within the game and is the center to which the playerChar, Enemy, and Item classes inherit from. It contains multiple private variables which detail the main components that comprise of every entity and object in the game being health, strength, defense as well as the x and y coordinates where each entity is placed upon the dungeon map detailed in the Floor class. It contains a default constructor as well as parameterized constructors as a fallback/inherited constructor to which all other entities can pull from which sets said basic stats and position information. The class also features multiple basic setters, getters, and modifiers which allow the user to easily modify and access the attributes of each entity as to track them throughout gameplay. In addition, it also features basic functions that each entity will be able to perform such as attacking another entity.

The PlayerChar class is an inherited class of the base class Entity. It similarly inherits the object attributes of health, strength, defense, as well as x and y coordinates. The player class is unique among other entities because it has the ability to be moved via player input using the movePlayer function in the Floor class as well as the fact it is capable of interacting with the environment, prompting menus for subsequent actions when it comes across any obstacle or another entity, say one of the Enemy class. It is also unique as it is capable of acquiring stat bonuses through the Item class and using its useItem function accordingly.

The Floor class is a class that is similar to the Menu class as it prints out the dungeon map directly into the terminal. It takes a random file input from a set of premade dungeon map layouts and uses a separate function to parse that file into a readable 2d array/vector to which we store the positions and coordinates of xy coords as well as the printing of each map layout. In addition, the Floor class houses the move function that all entities are capable of calling as it directly impacts the file every time it needs to be reprinted into the terminal. There, it is capable of checking whether or not an entity has made contact to another entity or wall to which it will promptly disable movement in that direction and prompt another response. For example, should the player use the wasd keys to move against a hashtag (wall), the program will stop the action and prompt the player that there is an object in the way, otherwise, the array will be updated as prescribed and the individual coordinates of each entity are updated and the new map is printed out.

The Enemy class is an inherited class of the base class Entity. It inherits the object into types of health, strength, defense, alongside  x and y coordinates. The Enemy class is meant to store the enemies objects in the game which includes snakes, zombies, and dragons. Each enemy type is an inherited class that inherits the Enemy class, and the entity class. Within the snake, zombie and dragon class, each constructor sets the level, health, strength and defense of each monster, alongside its coordinates. Moreover, within the constructor, bool alive is defined as true, however, is changed to false once the health of the monster enemy reaches zero. The function IsAlive, determines whether the monster type is alive or not. If health reaches zero, then the alive object is changed to false. Enemy types are stored in a vector, which all hold the separate status of the enemies. Each enemy type gets a vector, allowing for easier organization. Each enemy has a level, which modifies the preset statistics of its character. For example, a level 10 snake will additively do more damage than a level 1 snake. This modification of predetermined character statistics is done through the “lvl” functions which modify the effect enemies have on the player user with correlation to their level. Finally the Enemy class features a deconstructor, which will thoroughly delete all enemy and empty vector item’s types stored in the vector. 

Moreover, the item class is an inheritance class from both enemies and entities, and has two separate vectors (items, potions)  allowing for easier organization when it comes to the two different types of items. Momentary potions allow for the player's health or experience to grow. The other type of item inherited from the item class, are “items” or permanent increases to strength or defense. However, both potions or items share the same class functions. The level type object in one of the item inheritance classes, allows for the modification of the object's properties. For example, a level 1 health potion will restore significantly less health in comparison to a level 10 potion. This modification is done through the “lvl” functions, which modify the effects of what each item does to the player in correlation to the level of the item.
Finally the Enemy class features a deconstructor, which will thoroughly delete all enemy types stored in the vector. Finally the Item class features a deconstructor, which will thoroughly delete all item types stored in the vector.